Dropdown funcionando con las 3 opciones
